app/requirements.txt
numpy>=1.21.0,<1.27.0
streamlit>=1.28.0
tensorflow==2.15.0
Pillow>=9.0.0
h5py>=3.8.0